requirements.txt
annotated-types==0.7.0
black==24.10.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
docstring_parser==0.16
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-firestore==2.11.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.10.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.68.1
grpcio-status==1.62.3
html2text==2024.2.26
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
mypy-extensions==1.0.0
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
shapely==2.0.6
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
